Deleted old class, put wet code into extensions, rearranged folders a bit
