--- Lynked/UpdateServiceWebViewViewController.swift
@@ -1,24 +1,0 @@
-//
-//  UpdateServiceWebViewViewController.swift
-//  Lynked
-//
-//  Created by Matthew Howes Singleton on 6/13/17.
-//  Copyright © 2017 Matthew Singleton. All rights reserved.
-//
-
-import UIKit
-
-class UpdateServiceWebViewViewController: UIViewController {
-
-    @IBOutlet weak var webView: UIWebView!
-    
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-
-
-
-}
